issue: 1259552 Optimize get buffers procedure
Fix Jenkins

+++ src/vma/dev/cq_mgr.cpp
@@ -586,7 +586,7 @@ bool cq_mgr::compensate_qp_poll_success(mem_buf_desc_t* buff_cur)
 #endif // DEFINED_SOCKETXTREME
 		
 		if (m_rx_pool.size() || request_more_buffers()) {
-			m_qp_rec.debt -= m_qp_rec.qp->post_recv_buffers(&m_rx_pool, MIN(m_qp_rec.debt, m_rx_pool.size()));
+			m_qp_rec.debt -= m_qp_rec.qp->post_recv_buffers(&m_rx_pool, MIN((size_t)m_qp_rec.debt, m_rx_pool.size()));
 			m_p_cq_stat->n_buffer_pool_len = m_rx_pool.size();
 		}
 		else if (m_b_sysvar_cq_keep_qp_full ||
