+++ src/vma/dev/cq_mgr.inl
@@ -58,7 +58,7 @@ inline void cq_mgr::compensate_qp_poll_failed()
 	// Compensate QP for all completions debt
 	if (m_qp_rec.debt) {
 		if (likely(m_rx_pool.size() || request_more_buffers())) {
-			m_qp_rec.debt -= m_qp_rec.qp->post_recv_buffers(&m_rx_pool, MIN(m_qp_rec.debt, m_rx_pool.size()));
+			m_qp_rec.debt -= m_qp_rec.qp->post_recv_buffers(&m_rx_pool, MIN((size_t)m_qp_rec.debt, m_rx_pool.size()));
 			m_p_cq_stat->n_buffer_pool_len = m_rx_pool.size();
 		}
 	}
